Update README.md
Note: Some common Welsh words, fel 'oes', 'maen' and 'rwyf' are ambiguous and can refert to objects as wel as represent common verbal forms. We believe that it is appropriate that these forms are included, but advise that the stopword list be used carefully and that apprpriate changes are made according to the required use case.